Help channels: move name and channel funcs to separate modules
